+++ resources/js/components/ConsolePanel.js
@@ -4,10 +4,13 @@ export default class ConsolePanel
     consoleAutoScrollEnabled = true;
      
     state;
-   
+    firstRun;
+
     constructor(state)
     {
         this.state = state;
+        this.firstRun = true;
+
         console.log("Console panel", "constructor");
 
         window.addEventListener("message", (event) =>
@@ -20,40 +23,42 @@ export default class ConsolePanel
 
             if(event.data.message === "console-output")
             {
-                let consoleContainer = document.querySelector("#console-panel");
-                consoleContainer.innerHTML += `<div>${event.data.data}</div>`;
+                this.getElement().innerHTML += `<div>${event.data.data}</div>`;
                 
                 // auto scroll
                 if(this.consoleAutoScrollEnabled)
-                    consoleContainer.scrollTop = consoleContainer.scrollHeight;
+                    this.getElement().scrollTop = this.getElement().scrollHeight;
 
-                this.state.setActiveTab("Console");
+                if(this.firstRun)
+                {
+                    this.state.setActiveTab("console");
+                    this.firstRun = false;
+                }
             }
         });
     }
     
     clear()
     {
-        document.querySelector("#console-panel").innerHTML = "";
+        this.getElement().innerHTML = "";
     }
 
-    exists()
+    getElement()
     {
-        return this.consolePanelExist;
+        return document.querySelector('#console-panel');
     }
 
     onInit()
     {
-        let consoleContainer = document.querySelector("#console-panel");
         document.querySelector("#console-auto-scroll").addEventListener("click", () =>
         {
             this.consoleAutoScrollEnabled = true;
             document.querySelector("#console-auto-scroll").classList.toggle("hidden", this.consoleAutoScrollEnabled);
         });
 
-        consoleContainer.addEventListener("wheel", (event) =>
+        this.getElement().addEventListener("wheel", (event) =>
         {
-            let nearBottom = ((consoleContainer.scrollHeight - consoleContainer.clientHeight) <= (consoleContainer.scrollTop + 1));
+            let nearBottom = ((this.getElement().scrollHeight - this.getElement().clientHeight) <= (this.getElement().scrollTop + 1));
 
             if(nearBottom)
             {
@@ -61,7 +66,7 @@ export default class ConsolePanel
                 if(event.deltaY < 0)
                 {
                     this.consoleAutoScrollEnabled = false;
-                    consoleContainer.scrollTop = consoleContainer.scrollHeight - 20;
+                    this.getElement().scrollTop = this.getElement().scrollHeight - 20;
                     document.querySelector("#console-auto-scroll").classList.toggle("hidden", this.consoleAutoScrollEnabled);
                 }
             }
@@ -79,4 +84,8 @@ export default class ConsolePanel
         });
     }
 
+    setFirstRun()
+    {
+        this.firstRun = true;
+    }
 }